mmap04.c: Avoid vma merging
We hit a scenario where new mapping was merged with existing mapping of
same permission and the return address from the mmap was hidden in the
merged mapping in /proc/self/maps, causing the test to fail.
To avoid this, we first  create a 2-page mapping with the different
permissions, and then remap the 2nd page with the perms being tested.

Signed-off-by: Avinesh Kumar <akumar@suse.de>
Reported-by: Martin Doucha <mdoucha@suse.cz>
Signed-off-by: Avinesh Kumar <akumar@suse.de>